+++ CHANGELOG.rst
@@ -25,6 +25,11 @@ Added
   make regions with high intensity more transparent (!311).
 * New ``--index`` command-line option for ``volume`` overlays, allowing
   the indices for all non-spatial dimensions to be specified (!304).
+* New option to display the coordinates for the current location on the
+  canvases of an ortho view (available on the command-line via
+  ``--showLocation``) (!314).
+* New option to control the location cursor width on ortho/lightbox views
+  (available on the command-line via ``--cursorWidth``) (!314).
 
 
 Changed